refactor webpack config to support pixi
