+++ src/index.html
@@ -4,16 +4,19 @@
 	<meta charset="utf-8">
 	<meta http-equiv="X-UA-Compatible" content="IE=edge">
 	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<% if (process.env.NODE_ENV==="production" ) { %>
+	<% if (process.env.NODE_ENV === "production" ) { %>
 		<meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline' 'unsafe-eval'; style-src 'self' 'unsafe-inline' https://fonts.googleapis.com/;  img-src 'self' ">
 	<% } %>
-	<title>JSRL</title>
+	<title><%= htmlWebpackPlugin.options.title %></title>
 	<link href="./unicodetiles/unicodetiles.css" rel="stylesheet" type="text/css" />
 	<script src="./unicodetiles/unicodetiles.js"></script>
 	<script src="./unicodetiles/input.js"></script>
+
+	<% if (htmlWebpackPlugin.options.gfx === 'default') { %>
 	<script src="./unicodetiles/ut.WebGLRenderer.js"></script>
 	<script src="./unicodetiles/ut.CanvasRenderer.js"></script>
 	<script src="./unicodetiles/ut.DOMRenderer.js"></script>
+	<% } %>
 	<link rel="icon" href="./favicon.ico" sizes="any" />
 	<base href="./" />
 </head>
@@ -22,7 +25,11 @@
 		<div id="game"></div>
 	</div>
 	<div style="opacity: 0; position: absolute; top: 0; left: 0;">
-		<span style="font-family: DejaVuSansMono_ut;">-</span>
+		<% if (htmlWebpackPlugin.options.gfx === 'default') { %>
+			<span style="font-family: DejaVuSansMono_ut;">-</span>
+		<% } else { %>
+			<span style="font-family: 'Kenney Pixel';">-</span>
+		<% } %>
 	</div>
 </body>
 </html>
