+++ src/main.js
@@ -36,6 +36,8 @@ function createWindow () {
     mainWindow.loadFile('index.html');
   }
 
+  mainWindow.webContents.openDevTools();
+
 }
 
 // This method will be called when Electron has finished
