+++ webpack/webpack.config.js
@@ -1,75 +1,34 @@
-const path = require('path');
-const webpackTargetElectronRenderer = require('webpack-target-electron-renderer');
-var argv = require('minimist')(process.argv.slice(2));
-const isWeb = (argv && argv.target === 'web');
-const outputPath = (isWeb ? 'dist/web' : 'dist/electron');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const CopyWebpackPlugin = require("copy-webpack-plugin");
-const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 
-let options = {
-  module: {
-    rules: [
-      {
-        test: /\.js(x)?$/,
-        exclude: /node_modules/,
-        use: ['babel-loader']
-      },
-      {
-        test: /\.css$/i,
-        use: ['style-loader', 'css-loader'],
-      },
-    ],
-  },
+require('dotenv').config()
+const path = require('path')
+const { merge } = require('webpack-merge')
+const commonConfig = require('./webpack.base.config.js')
+const defaultConfig = require('./webpack.default.config.js')
+const gfxConfig = require('./webpack.gfx.config.js')
+const HtmlWebpackPlugin = require('html-webpack-plugin')
 
-  output: {
-    path: path.resolve(__dirname, '..', outputPath),
-    filename: '[name]_bundle.js',
-  },
+module.exports = (env, args) => {
 
-  resolve: {
-    extensions: ['', '.js', '.jsx'],
-  },
+  const mode = args.mode
+  const target = process.env.target || 'web'
+  const gfx = process.env.gfx || 'default'
+  let config
 
-  entry: {
-    renderer: path.resolve(__dirname, '..', './src/renderer.js')
-  },
-
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: 'src/index.html',
-      inject: true
-    }),
-
-    new CopyWebpackPlugin({
-      patterns: [
-        { from: path.resolve(__dirname, '..', 'public'), to: path.resolve(__dirname, '..', outputPath) },
-      ],
-    }),
-
-    new CleanWebpackPlugin(),
-  ],
-
-  devServer: {
-    static: {
-      directory: path.join(__dirname, '..', 'public'),
-    },
-    watchFiles: ['src/index.html'],
+  if (gfx !== 'default') {
+    config = merge(commonConfig, gfxConfig)
+  } else {
+    config = merge(commonConfig, defaultConfig)
   }
 
-};
+  config.plugins.push(
+    new HtmlWebpackPlugin({
+      template: 'src/index.html',
+      inject: true,
+      title: `JSRL | ${gfx} | ${target}`,
+      gfx: gfx
+    })
+  )
 
-options.target = webpackTargetElectronRenderer(options);
 
-if (!isWeb) {
-  options.plugins.push(new CopyWebpackPlugin({
-    patterns: [
-      {from: path.resolve(__dirname, '..', 'src/package.json'), to: path.resolve(__dirname, '..', outputPath) },
-      {from: path.resolve(__dirname, '..', 'src/main.js'), to: path.resolve(__dirname, '..', outputPath) },
-      {from: path.resolve(__dirname, '..', 'src/preload.js'), to: path.resolve(__dirname, '..', outputPath) },
-    ]
-  }))
+  return config
 }
-
-
-module.exports = options;
