+++ webpack/webpack.default.config.js
@@ -3,8 +3,9 @@ const {resolve} = require('path')
 const options = {
 
   entry: {
-    renderer: resolve(__dirname, '..', './src/renderer.gfx.js')
+    renderer: resolve(__dirname, '..', './src/renderer.js')
   },
+
 }
 
 module.exports = options
